+++ react/store/ReturnRequest/StoreReturnDetails.tsx
@@ -17,6 +17,9 @@ import { StatusTimeline } from '../../common/components/ReturnDetails/StatusTime
 import { StatusHistory } from '../../common/components/ReturnDetails/StatusHistory'
 import { OrderLink } from '../../common/components/ReturnDetails/OrderLink'
 import { CurrentRequestStatus } from '../../common/components/ReturnDetails/CurrentRequestStatus'
+import RequestCancellation from '../../common/components/ReturnDetails/RequestCancellation'
+import { UpdateRequestStatusProvider } from '../../admin/provider/UpdateRequestStatusProvider'
+import { AlertProvider } from '../../admin/provider/AlertProvider'
 
 const CSS_HANDLES = ['contactPickupContainer'] as const
 
@@ -26,9 +29,8 @@ const StoreReturnDetails = () => {
   const handles = useCssHandles(CSS_HANDLES)
 
   return (
-    <PageBlock className="ph0 mh0 pa0 pa0-ns">
+    <PageBlock>
       <PageHeader
-        className="ph0 mh0 nl5"
         title={
           <FormattedMessage id="store/return-app.return-request-details.page-header.title" />
         }
@@ -40,7 +42,14 @@ const StoreReturnDetails = () => {
             to: '#/my-returns',
           })
         }}
-      />
+      >
+        <AlertProvider>
+          <UpdateRequestStatusProvider>
+            <RequestCancellation />
+          </UpdateRequestStatusProvider>
+        </AlertProvider>
+      </PageHeader>
+
       <StoreReturnDetailsLoader data={{ loading, error }}>
         <CurrentRequestStatus />
         <OrderLink />